% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbds.R
\docType{data}
\name{adbds}
\alias{adbds}
\title{adbds}
\format{
a data frame with 16,893 rows and 25 columns. Retrieved from
\url{https://github.com/RhoInc/data-library/tree/master/data/clinical-trials/renderer-specific}.
}
\usage{
adbds
}
\description{
labs dataset in tall format that includes baseline
characteristics.
}
\keyword{datasets}
