% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{adjusted_ggplotly}
\alias{adjusted_ggplotly}
\title{Adjust and Convert ggplot to Plotly}
\usage{
adjusted_ggplotly(p)
}
\arguments{
\item{p}{A ggplot object to be adjusted and converted.}
}
\value{
A plotly object with adjusted layout settings.
}
\description{
This function converts a ggplot object to a plotly object and adjusts the layout to ensure consistent spacing
for the y-axis labels and positions the legend below the plot.
}
