% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{adlb}
\format{
a data frame with 13708 rows and 26 columns. Retrieved from
\url{https://github.com/RhoInc/data-library/tree/master/data/clinical-trials/renderer-specific}
and excluded VS tests to keep only lab tests. BLFL also derived.
}
\usage{
adlb
}
\description{
labs dataset in tall format that includes baseline
characteristics.
}
\keyword{datasets}
