% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{adsl}
\format{
a data frame with 150 rows and 14 columns. Retrieved from
  \url{https://github.com/RhoInc/data-library/tree/master/data/clinical-trials/renderer-specific} and
  selected adsl variables and removed duplicates. BLFL also derived.
}
\usage{
adsl
}
\description{
subject-level dataset that includes demographic information.
}
\keyword{datasets}
