% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_checks_util.R
\name{check_expected_cols}
\alias{check_expected_cols}
\title{Check for Missing Columns in Data}
\usage{
check_expected_cols(data_domain, settings_list, data_list)
}
\arguments{
\item{data_domain}{A string specifying the data domain to check (e.g., "labs", "vitals", "dm").}

\item{settings_list}{A list containing the settings for each data domain.}

\item{data_list}{A list containing the data for each domain.}
}
\value{
A string message listing the missing columns, or NULL if all expected columns are present.
}
\description{
This function checks for any columns specified in the settings object that are not present in the data.
It returns a message listing the missing columns, if any.
}
