% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_checks_util.R
\name{check_required_settings}
\alias{check_required_settings}
\title{Check Required Settings in a List}
\usage{
check_required_settings(required_list, actual_list, parent_key = "")
}
\arguments{
\item{required_list}{A list specifying the required elements.}

\item{actual_list}{A list to be checked against the required elements.}

\item{parent_key}{(Optional) A string used internally to track the full key path during recursion.
Default is an empty string.}
}
\value{
A list of missing elements, if any.
}
\description{
This function checks if a given list contains all the required elements specified in another list.
It allows for additional elements at any nesting level and prints out a message listing the missing settings, if any.
}
