% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nepexplorer_app.R
\name{create_nepexplorer_app}
\alias{create_nepexplorer_app}
\title{Run the core NepExplorer App}
\usage{
create_nepexplorer_app(
  data = list(dm = nepExplorer::adsl, labs = nepExplorer::adlb, vitals =
    nepExplorer::advs),
  mapping = NULL,
  runNow = TRUE
)
}
\arguments{
\item{data}{datasets to be loaded in to the app. Sample nepExplorer data used
by default}

\item{mapping}{list specifying the initial mapping values for each data
mapping for each domain (e.g. list(aes= list(id_col='USUBJID',
seq_col='AESEQ')).}

\item{runNow}{Should the shiny app object created be run directly? Helpful
when writing  functions to dispatch to shinyapps, rsconnect, or shinyproxy.}
}
\description{
Run the core NepExplorer App
}
