% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_data_fcn.R
\name{creatinine_data_fcn}
\alias{creatinine_data_fcn}
\title{Create Creatinine Level Data and Patient Level Stages}
\usage{
creatinine_data_fcn(df, settings)
}
\arguments{
\item{df}{lab dataset in tall format with creatinine lab}

\item{settings}{settings object with column mappings}
}
\description{
Create Creatinine Level Data and Patient Level Stages
}
