% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_mod.R
\name{creatinineScatterUI}
\alias{creatinineScatterUI}
\title{Creatine Scatter Plot + Summary Table UI}
\usage{
creatinineScatterUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Creatine Scatter Plot + Summary Table UI
}
