% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_charts.R
\name{draw_summary_table}
\alias{draw_summary_table}
\title{Draw Creatinine Summary Table}
\usage{
draw_summary_table(df)
}
\arguments{
\item{df}{lab dataset in tall format with creatinine lab}
}
\description{
Draw Creatinine Summary Table
}
