% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{drawDemoTable}
\alias{drawDemoTable}
\title{Draw Demography Table}
\usage{
drawDemoTable(
  adlb,
  settings,
  demo_vars = c("USUBJID", "AGE", "SEX", "RACE", "ARM")
)
}
\arguments{
\item{adlb}{lab data in tall format that must contain DY for study day,
VISITN for visit number, TEST for lab test, and STRESN for lab value}

\item{settings}{settings object with column mappings}

\item{demo_vars}{character vector of column names to include in demography
table - those not found in data will be ignored}
}
\value{
gt object
}
\description{
Draw Demography Table
}
