% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{drawNormalizedAlbumin}
\alias{drawNormalizedAlbumin}
\title{Draw Time Series Plot for Normalized Albumin}
\usage{
drawNormalizedAlbumin(adlb)
}
\arguments{
\item{adlb}{lab data in tall format that must contain DY for study day,
VISITN for visit number, TEST for lab test, and STRESN for lab value}
}
\value{
ggplot object
}
\description{
Draw Time Series Plot for Normalized Albumin
}
