% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{drawPercentChange}
\alias{drawPercentChange}
\title{drawPercentChange}
\usage{
drawPercentChange(
  adlb,
  labs = c("Creatinine", "Cystatin C"),
  KDIGO_reference_ranges = TRUE
)
}
\arguments{
\item{adlb}{lab data in long format that must contain DY for study day, VISITN for visit number, TEST for lab test, and STRESN for lab value}

\item{labs}{character string or character vector specifying which labs from TEST to include}
}
\value{
ggplot object
}
\description{
Draw Time Series Plot for Percent Change in Lab Values from Baseline (add dependencies like pipe later)
}
