% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{drawRawChange}
\alias{drawRawChange}
\title{Draw Time Series Plot for Raw Change in Lab Values from Baseline}
\usage{
drawRawChange(
  adlb,
  labs = c("Creatinine", "Cystatin C"),
  delta_creatinine_ref_ranges = TRUE
)
}
\arguments{
\item{adlb}{lab data in tall format that must contain DY for study day,
VISITN for visit number, TEST for lab test, and STRESN for lab value}

\item{labs}{character string or character vector specifying which labs from
TEST to include}

\item{delta_creatinine_ref_ranges}{boolean for whether or not to
include horizontal lines denoting delta creatinine reference ranges}
}
\value{
ggplot object
}
\description{
Draw Time Series Plot for Raw Change in Lab Values from Baseline
}
