% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_charts.R
\name{drawULNFoldChange}
\alias{drawULNFoldChange}
\title{Draw Time Series Plot for Fold Change from ULN}
\usage{
drawULNFoldChange(
  adlb,
  labs = c("Bicarbonate", "Blood Urea Nitrogen", "Calcium", "Chloride", "Phosphorus",
    "Potassium", "Sodium")
)
}
\arguments{
\item{adlb}{lab data in tall format that must contain DY for study day, VISITN for visit number, TEST for lab test, and STRESN for lab value}

\item{labs}{character string or character vector specifying which labs from TEST to include}
}
\value{
ggplot object
}
\description{
Draw Time Series Plot for Fold Change from ULN
}
