% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_nepExplorer.R
\name{init_nepExplorer}
\alias{init_nepExplorer}
\title{Initialize Settings for nepExplorer}
\usage{
init_nepExplorer(data, settings)
}
\arguments{
\item{data}{`list` Named list of data frames that includes participant-level subject data (`dm`)
and event-level adverse event data (`aes`).}

\item{settings}{`list` Named list of settings.}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for nepExplorer
}
