% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepexplorerMod.R
\name{nepexplorer_server}
\alias{nepexplorer_server}
\title{Nep Explorer Module - Server}
\usage{
nepexplorer_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with \code{data} and \code{settings} options.}
}
\value{
returns shiny module Server function
}
\description{
Nep Explorer Module - Server
}
