% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepexplorerMod.R
\name{nepexplorer_ui}
\alias{nepexplorer_ui}
\title{Nep Explorer Module - UI}
\usage{
nepexplorer_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Nep Explorer Module - UI
}
