% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_mod.R
\name{patientProfileServer}
\alias{patientProfileServer}
\title{Patient Profile Module Server}
\usage{
patientProfileServer(id, df, settings, subj_id, patient_measures)
}
\arguments{
\item{id}{module id}

\item{df}{lab dataset in tall format with creatinine lab}

\item{settings}{settings object with column mappings}

\item{subj_id}{single subject ID as character string}
}
\value{
returns shiny server module
}
\description{
Patient Profile Module Server
}
