% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_profile_mod.R
\name{patientProfileUI}
\alias{patientProfileUI}
\title{Patient Profile Module UI}
\usage{
patientProfileUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Patient Profile Module UI
}
