% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_mod.R
\name{creatinineScatterServer}
\alias{creatinineScatterServer}
\title{Creatine Scatter Plot + Summary Table UI}
\usage{
creatinineScatterServer(
  id,
  df,
  settings,
  animate,
  animation_transition_time,
  animation_time_unit
)
}
\arguments{
\item{id}{module id}

\item{df}{lab dataset in tall format with creatinine lab}

\item{settings}{settings object with column mappings}

\item{animate}{"off" or "on", "on" displays time animation}

\item{animation_transition_time}{frame transition speed in seconds}

\item{animation_time_unit}{column for study day or visit to be used as time variable for animation}
}
\value{
returns shiny server module
}
\description{
Creatine Scatter Plot + Summary Table UI
}
