% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_mod.R
\name{creatinineScatterServer}
\alias{creatinineScatterServer}
\title{Creatine Scatter Plot + Summary Table UI}
\usage{
creatinineScatterServer(id, df, settings, animate)
}
\arguments{
\item{id}{module id}

\item{df}{lab dataset in tall format with creatinine lab}

\item{settings}{settings object with column mappings}
}
\value{
returns shiny server module
}
\description{
Creatine Scatter Plot + Summary Table UI
}
