% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_charts.R
\name{draw_creatinine_scatter}
\alias{draw_creatinine_scatter}
\title{Draw Creatinine Scatter Plot}
\usage{
draw_creatinine_scatter(
  df,
  settings,
  animate = "off",
  animation_transition_time = 0.5,
  animation_time_unit
)
}
\arguments{
\item{df}{lab dataset in tall format with creatinine lab}

\item{settings}{settings object with column mappings}

\item{animate}{"off" or "on", "on" displays time animation}

\item{animation_transition_time}{frame transition speed in seconds}

\item{animation_time_unit}{column for study day or visit to be used as time variable for animation}
}
\description{
Draw Creatinine Scatter Plot
}
