% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatinine_scatter_charts.R
\name{draw_creatinine_scatter}
\alias{draw_creatinine_scatter}
\title{Draw Creatinine Scatter Plot}
\usage{
draw_creatinine_scatter(df)
}
\arguments{
\item{df}{lab dataset in tall format with creatinine lab}
}
\description{
Draw Creatinine Scatter Plot
}
