% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_nepExplorer.R
\name{init_nepExplorer}
\alias{init_nepExplorer}
\title{Initialize Settings for nepExplorer}
\usage{
init_nepExplorer(data, settings)
}
\arguments{
\item{data}{\code{list} Named list of data frames that includes participant-level subject data (\code{dm})
labs(\code{labs}) and vital signs data (\code{vitals}).}

\item{settings}{\code{list} Named list of settings.}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for nepExplorer
}
